refactor: remove and add some t function, change ErrorCode translation, etc #1
