+++ packages/ui/public/locales/en/translation.json
@@ -64,17 +64,18 @@
   "Create your first account now!": "",
   "Click to copy address": "",
   "Address copied!": "",
-  "UnknownRequest": "Unknown Request",
-  "UnknownRequestOrigin": "Unknown Request Origin",
-  "InvalidMessageFormat": "Invalid Message Format",
-  "KeypairNotFound": "Keypair Not Found",
-  "AccountNotFound": "Account Not Found",
-  "KeyringNotInitialized": "Keyring Not Initialized",
-  "PasswordIncorrect": "Password Incorrect",
-  "PasswordRequired": "Password Required",
-  "WalletLocked": "Wallet Locked",
-  "KeyringLocked": "Keyring Locked",
-  "AccountNameRequired": "Account Name Required",
-  "AccountNameUsed": "Account Name Used",
-  "InternalError": "Internal Error"
+  "UnknownRequest": "Unknown request",
+  "UnknownRequestOrigin": "Unknown request origin",
+  "InvalidMessageFormat": "Invalid message format",
+  "KeypairNotFound": "Keypair not found",
+  "AccountNotFound": "Account not found",
+  "KeyringNotInitialized": "Keyring is not initialized",
+  "PasswordIncorrect": "Password incorrect",
+  "PasswordRequired": "Password required",
+  "WalletLocked": "The wallet is locked, please unlock it first",
+  "KeyringLocked": "The keyring is locked, please unlock the wallet first",
+  "AccountNameRequired": "Account name is required",
+  "AccountNameUsed": "Account name is already picked",
+  "InternalError": "Internal error",
+  "Invalid request": ""
 }
