+++ packages/ui/src/components/pages/Request/InvalidRequest.tsx
@@ -7,8 +7,8 @@ const InvalidRequest: FC<Props> = ({ className = '', reason }) => {
 
   return (
     <div className={`${className} text-center`}>
-      <h2>{t<string>('Invalid Request')}</h2>
-      {reason && <p>Reason: {reason}</p>}
+      <h2>{t<string>('Invalid request')}</h2>
+      {reason && <p>Reason: {t<string>(reason)}</p>}
       <p>{t<string>("If you open this page by accident, it's safe to close it now.")}</p>
     </div>
   );
