+++ packages/ui/src/components/shared/__tests__/SettingsWalletButton.spec.tsx
@@ -1,5 +1,5 @@
 import { initializeKeyring, newUser, render, screen, UserEvent, waitFor } from '__tests__/testUtils';
-import SettingsWalletButton from '../SettingsWalletButton';
+import SettingsWalletButton from '../settings/SettingsWalletButton';
 
 describe('SettingsWalletButton', () => {
   it('should hide the dialog by default', async () => {
